refactor(tests): rename assertFailure to assert.failsAsync
